+++ pdf2mind.py
@@ -1,8 +1,7 @@
-
-from parser import cmd_parser 
+from utils.parser import cmd_parser 
 from pdf.pdf_reader import PdfProcess
 from mind.generator import gen_all_mindmaps
-from utils import model_selector, format_selector
+from utils.utils import model_selector, format_selector
 
 def main():
     args = cmd_parser()
